Show hot path on presenter
- Flash hot path every 6 seconds
- Also use hot path to calculate game win
